+++ CHANGELOG.md
@@ -6,10 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Settings options are not searchable
+
 ## [0.1.2]
 ### Fixed
 - Completion results on Windows no longer show full path [#11](https://github.com/hailwood/idea-inertiajs-plugin/issues/11)
-
+- 
 ## [0.1.1]
 ### Added
 - Notification when Inertia root does not exist